Issue-413: Setup for travis-ci
* No code changes, just renaming.
* Removed __init__.py files from tests directory because they were causing problems with IntelliJ unit test runner.
* Added blank test cases for classes that have no test yet.
