+++ .coveragerc
@@ -1,0 +1,6 @@
+[run]
+omit =
+    */__init__.py
+    setup.py
+    test-setup.py
+    tests
