+++ requirements.txt
@@ -1,3 +1,4 @@
+mock
 boto3
 gogs_client
 bs4
