+++ tests/aws_tools/test_s3_handler.py
@@ -1,0 +1,6 @@
+from __future__ import unicode_literals
+import unittest
+
+
+class S3HandlerTests(unittest.TestCase):
+    pass
