+++ tests/client_tools/test_client_tools.py
@@ -1,0 +1,6 @@
+from __future__ import unicode_literals
+import unittest
+
+
+class ClientToolsTests(unittest.TestCase):
+    pass
