+++ tests/door43_tools/test_language_handler.py
@@ -1,0 +1,6 @@
+from __future__ import unicode_literals
+import unittest
+
+
+class LanguageHandlerTests(unittest.TestCase):
+    pass
