+++ tests/door43_tools/test_manifest_handler.py
@@ -1,0 +1,6 @@
+from __future__ import unicode_literals
+import unittest
+
+
+class ManifestHandlerTests(unittest.TestCase):
+    pass
