--- tests/gogs_tools/__init__.py
@@ -1,1 +1,0 @@
-import test
